# Performance improvements in the PNG filtering (roughly 20% gain on my machine).