--- a/src/ImageSharp/Formats/Png/Filters/PaethFilter.cs
+++ b/src/ImageSharp/Formats/Png/Filters/PaethFilter.cs
@@ -44,7 +44,7 @@ namespace SixLabors.ImageSharp.Formats.Png.Filters
                 byte left = Unsafe.Add(ref scanBaseRef, x - bytesPerPixel);
                 byte above = Unsafe.Add(ref prevBaseRef, x);
                 byte upperLeft = Unsafe.Add(ref prevBaseRef, x - bytesPerPixel);
-                scan = (byte)(scan + PaethPredicator(left, above, upperLeft));
+                scan = (byte)(scan + PaethPredictor(left, above, upperLeft));
             }
         }
 
@@ -70,6 +70,7 @@ namespace SixLabors.ImageSharp.Formats.Png.Filters
             // Paeth(x) = Raw(x) - PaethPredictor(Raw(x-bpp), Prior(x), Prior(x - bpp))
             resultBaseRef = 4;
 
+#if OLD_AND_SLOW
             for (int x = 0; x < scanline.Length; x++)
             {
                 if (x - bytesPerPixel < 0)
@@ -91,6 +92,27 @@ namespace SixLabors.ImageSharp.Formats.Png.Filters
                     sum += res < 128 ? res : 256 - res;
                 }
             }
+#else
+            int x = 0;
+            for (; x < bytesPerPixel; ++x) {
+                byte scan = Unsafe.Add(ref scanBaseRef, x);
+                byte above = Unsafe.Add(ref prevBaseRef, x);
+                ref byte res = ref Unsafe.Add(ref resultBaseRef, x + 1);
+                res = (byte)(scan - PaethPredictor(0, above, 0));
+                sum += ImageMaths.FastAbs(unchecked((sbyte)res));
+            }
+
+            for (int xLeft = x - bytesPerPixel; x < scanline.Length; ++xLeft) {
+                byte scan = Unsafe.Add(ref scanBaseRef, x);
+                byte left = Unsafe.Add(ref scanBaseRef, xLeft);
+                byte above = Unsafe.Add(ref prevBaseRef, x);
+                byte upperLeft = Unsafe.Add(ref prevBaseRef, xLeft);
+                ++x;
+                ref byte res = ref Unsafe.Add(ref resultBaseRef, x);
+                res = (byte)(scan - PaethPredictor(left, above, upperLeft));
+                sum += ImageMaths.FastAbs(unchecked((sbyte)res));
+            }
+#endif
 
             sum -= 4;
         }
@@ -106,7 +128,7 @@ namespace SixLabors.ImageSharp.Formats.Png.Filters
         /// The <see cref="byte"/>.
         /// </returns>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static byte PaethPredicator(byte left, byte above, byte upperLeft)
+        private static byte PaethPredictor(byte left, byte above, byte upperLeft)
         {
             int p = left + above - upperLeft;
             int pa = ImageMaths.FastAbs(p - left);