Ditched the use of modulo (%). Modulo is slow, and if you use it to just get the lower 8 bits (1 byte), you can just directly cast to byte.

Also moved one if < out of the loop for speed. The less branching the better.

Similarly replaced an if < 128 by cast to sbyte and taking its abs value.

Performance gain in writing to PNG file seems to be roughly 20% (release mode, 1000x1000 bitmap).

In ImageFramesCollectionTests I've set the culture to en-US to ensure getting english exception texts.

Edit: I've just done similar changes to the PNG decoding, but there's no or little performance improvement there.

EncodePng benchmark (before change):

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.371)
Intel Core i7-6500U CPU 2.50GHz (Skylake), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.1.104
  [Host]     : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-MJBKJC : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-MMODLS : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0

LaunchCount=1  TargetCount=3  WarmupCount=3  

```
|               Method | Runtime | LargeImage |     Mean |     Error |    StdDev | Scaled | ScaledSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|--------------------- |-------- |----------- |---------:|----------:|----------:|-------:|---------:|---------:|---------:|---------:|----------:|
| &#39;System.Drawing Png&#39; |    Core |      False | 16.98 ms |  3.807 ms | 0.2151 ms |   1.00 |     0.00 | 781.2500 | 218.7500 | 218.7500 |   2.03 MB |
|     &#39;ImageSharp Png&#39; |    Core |      False | 61.95 ms | 19.663 ms | 1.1110 ms |   3.65 |     0.07 | 250.0000 | 187.5000 | 187.5000 |   1.18 MB |
|                      |         |            |          |           |           |        |          |          |          |          |           |
| &#39;System.Drawing Png&#39; |     Clr |      False | 16.15 ms |  2.070 ms | 0.1169 ms |   1.00 |     0.00 | 781.2500 | 218.7500 | 218.7500 |   2.03 MB |
|     &#39;ImageSharp Png&#39; |     Clr |      False | 81.96 ms |  5.216 ms | 0.2947 ms |   5.07 |     0.03 | 250.0000 | 187.5000 | 187.5000 |   1.18 MB |

EncodePng benchmark (after change):

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.371)
Intel Core i7-6500U CPU 2.50GHz (Skylake), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.1.104
  [Host]     : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-ATQUUE : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-QUFUDX : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0

LaunchCount=1  TargetCount=3  WarmupCount=3  

```
|               Method | Runtime | LargeImage |     Mean |      Error |    StdDev | Scaled | ScaledSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|--------------------- |-------- |----------- |---------:|-----------:|----------:|-------:|---------:|---------:|---------:|---------:|----------:|
| &#39;System.Drawing Png&#39; |    Core |      False | 16.20 ms |  0.6100 ms | 0.0345 ms |   1.00 |     0.00 | 781.2500 | 218.7500 | 218.7500 |   2.03 MB |
|     &#39;ImageSharp Png&#39; |    Core |      False | 52.96 ms | 28.2241 ms | 1.5947 ms |   3.27 |     0.08 | 250.0000 | 187.5000 | 187.5000 |   1.18 MB |
|                      |         |            |          |            |           |        |          |          |          |          |           |
| &#39;System.Drawing Png&#39; |     Clr |      False | 16.45 ms |  3.3311 ms | 0.1882 ms |   1.00 |     0.00 | 781.2500 | 218.7500 | 218.7500 |   2.03 MB |
|     &#39;ImageSharp Png&#39; |     Clr |      False | 74.80 ms | 32.4062 ms | 1.8310 ms |   4.55 |     0.10 | 250.0000 | 187.5000 | 187.5000 |   1.18 MB |

DecodePng benchmark (before change):

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.371)
Intel Core i7-6500U CPU 2.50GHz (Skylake), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.1.104
  [Host]     : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-ADPASD : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-JVLCTB : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0

LaunchCount=1  TargetCount=3  WarmupCount=3  

```
|               Method | Runtime |      TestImage |     Mean |     Error |    StdDev | Scaled | ScaledSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------------- |-------- |--------------- |---------:|----------:|----------:|-------:|---------:|--------:|--------:|--------:|----------:|
| &#39;System.Drawing Png&#39; |    Core | Png/splash.png | 3.020 ms | 0.9129 ms | 0.0516 ms |   1.00 |     0.00 | 74.2188 | 74.2188 | 74.2188 | 239.95 KB |
|     &#39;ImageSharp Png&#39; |    Core | Png/splash.png | 4.021 ms | 2.3086 ms | 0.1304 ms |   1.33 |     0.04 |       - |       - |       - |   2.21 KB |
|                      |         |                |          |           |           |        |          |         |         |         |           |
| &#39;System.Drawing Png&#39; |     Clr | Png/splash.png | 2.922 ms | 0.5067 ms | 0.0286 ms |   1.00 |     0.00 | 74.2188 | 74.2188 | 74.2188 | 240.54 KB |
|     &#39;ImageSharp Png&#39; |     Clr | Png/splash.png | 7.546 ms | 4.3393 ms | 0.2452 ms |   2.58 |     0.07 | 46.8750 |       - |       - |  96.33 KB |

DecodePng benchmark (after change), no observable improvement here:

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.371)
Intel Core i7-6500U CPU 2.50GHz (Skylake), 1 CPU, 4 logical cores and 2 physical cores
Frequency=2531247 Hz, Resolution=395.0622 ns, Timer=TSC
.NET Core SDK=2.1.104
  [Host]     : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-PLYAIO : .NET Core 2.0.6 (Framework 4.6.26212.01), 64bit RyuJIT
  Job-WWJIGX : .NET Framework 4.6.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2633.0

LaunchCount=1  TargetCount=3  WarmupCount=3  

```
|               Method | Runtime |      TestImage |     Mean |     Error |    StdDev | Scaled | ScaledSD |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------------- |-------- |--------------- |---------:|----------:|----------:|-------:|---------:|--------:|--------:|--------:|----------:|
| &#39;System.Drawing Png&#39; |    Core | Png/splash.png | 2.991 ms | 1.0078 ms | 0.0569 ms |   1.00 |     0.00 | 74.2188 | 74.2188 | 74.2188 | 239.95 KB |
|     &#39;ImageSharp Png&#39; |    Core | Png/splash.png | 3.942 ms | 1.4060 ms | 0.0794 ms |   1.32 |     0.03 |       - |       - |       - |   2.21 KB |
|                      |         |                |          |           |           |        |          |         |         |         |           |
| &#39;System.Drawing Png&#39; |     Clr | Png/splash.png | 2.963 ms | 0.4208 ms | 0.0238 ms |   1.00 |     0.00 | 74.2188 | 74.2188 | 74.2188 | 240.54 KB |
|     &#39;ImageSharp Png&#39; |     Clr | Png/splash.png | 7.524 ms | 2.1313 ms | 0.1204 ms |   2.54 |     0.04 | 46.8750 |       - |       - |  96.33 KB |

### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [ ] I have provided test coverage for my change (where applicable)

### Description
<!-- A description of the changes proposed in the pull-request -->

<!-- Thanks for contributing to ImageSharp! -->